# backend/src/openapi.yaml
openapi: 3.0.1
info:
  title: Personal Finance Tracker API
  version: 0.1.0
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: ok
  /api/auth/signup:
    post:
      summary: Create new user account
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [USER, ADMIN, READ_ONLY]
      responses:
        '200':
          description: Account created successfully
  /api/categories:
    get:
      summary: Get all categories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of categories with amounts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    amount:
                      type: number
                      description: Initial/budget amount for the category
                    totalAmount:
                      type: number
                      description: Total amount spent in this category
                    transactionCount:
                      type: integer
                      description: Number of transactions in this category
                    createdAt:
                      type: string
                      format: date-time
                    updatedAt:
                      type: string
                      format: date-time
    post:
      summary: Create new category (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 50
                amount:
                  type: number
                  minimum: 0
                  description: Initial/budget amount for the category
              required:
                - name
      responses:
        '200':
          description: Category created successfully
  /api/categories/{id}:
    put:
      summary: Update category (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 50
                amount:
                  type: number
                  minimum: 0
      responses:
        '200':
          description: Category updated successfully
    delete:
      summary: Delete category (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category deleted successfully
  /api/transactions:
    get:
      summary: List transactions
      security:
        - bearerAuth: []
      responses:
        '200': { description: ok }